# Improve host to device memory copy