This PR removes CPU overhead.