@okuta 2018-06-19
Cloud you merge #1385 first?