@kmaehashi 2018-06-19
@okuta Could you check #1393 to fix jenkins failure?