@kmaehashi 2018-06-22
I benchmarked this PR with [this code](https://github.com/cupy/cupy-benchmark/pull/10) and confirmed performance improvement! 🎉 

Current master:

```
====== ========== ======== ======== ========
--              in_order / out_order
------ -------------------------------------
 ndim    C / C     C / F    F / C    F / F
====== ========== ======== ======== ========
  1      1.53ms    1.47ms   1.46ms   1.40ms
  2      1.44ms    1.47ms   1.51ms   1.43ms
  5      1.49ms    3.78ms   2.34ms   2.91ms
  8     750.61ms   4.62s    1.36s    4.04s
====== ========== ======== ======== ========
```

This PR:

```
====== ========== ======== ======== ==========
--               in_order / out_order         
------ ---------------------------------------
 ndim    C / C     C / F    F / C     F / F   
====== ========== ======== ======== ==========
  1      1.31ms    1.30ms   1.37ms    1.29ms  
  2      1.41ms    1.47ms   1.34ms    1.40ms  
  5      1.47ms    2.38ms   2.35ms    1.55ms  
  8     594.14ms   1.08s    2.39s    635.28ms 
====== ========== ======== ======== ==========
```