@okuta 2018-06-25
Thank you for your performance tests.
I fixed.